etl-jobs/requirements.txt
psycopg2-binary==2.9.7
elasticsearch==8.8.0